python/requirements.txt
discord.py
requests
mysql-connector-python
python-dotenv
